[BUG] Inventree fiddles with files directly rather than using Django Storage api Fixes #2585
