+++ InvenTree/InvenTree/config.py
@@ -3,8 +3,9 @@ Helper functions for loading InvenTree configuration options
 """
 
 import os
-import shutil
 import logging
+
+from django.core.files.storage import default_storage
 
 
 logger = logging.getLogger('inventree')
@@ -26,18 +27,15 @@ def get_config_file():
 
     cfg_filename = os.getenv('INVENTREE_CONFIG_FILE')
 
-    if cfg_filename:
-        cfg_filename = cfg_filename.strip()
-        cfg_filename = os.path.abspath(cfg_filename)
-    else:
+    if not cfg_filename:
         # Config file is *not* specified - use the default
-        cfg_filename = os.path.join(base_dir, 'config.yaml')
+        cfg_filename = 'config.yaml'
 
-    if not os.path.exists(cfg_filename):
+    if not default_storage.exists(cfg_filename):
         print("InvenTree configuration file 'config.yaml' not found - creating default file")
 
         cfg_template = os.path.join(base_dir, "config_template.yaml")
-        shutil.copyfile(cfg_template, cfg_filename)
+        default_storage.save(cfg_filename, open(cfg_template, 'rb').read())
         print(f"Created config file {cfg_filename}")
 
     return cfg_filename
@@ -53,18 +51,15 @@ def get_plugin_file():
     PLUGIN_FILE = os.getenv('INVENTREE_PLUGIN_FILE')
 
     if not PLUGIN_FILE:
-        # If not specified, look in the same directory as the configuration file
+        # If not specified, put into media root
+        PLUGIN_FILE = 'plugins.txt'
 
-        config_dir = os.path.dirname(get_config_file())
-
-        PLUGIN_FILE = os.path.join(config_dir, 'plugins.txt')
-
-    if not os.path.exists(PLUGIN_FILE):
+    if not default_storage.exists(PLUGIN_FILE):
         logger.warning("Plugin configuration file does not exist")
         logger.info(f"Creating plugin file at '{PLUGIN_FILE}'")
 
         # If opening the file fails (no write permission, for example), then this will throw an error
-        with open(PLUGIN_FILE, 'w') as plugin_file:
+        with default_storage.open(PLUGIN_FILE, 'w') as plugin_file:
             plugin_file.write("# InvenTree Plugins (uses PIP framework to install)\n\n")
 
     return PLUGIN_FILE
